Añadir validaciones, refactorizazción código

+++ src/main/java/com/turing/alan/cpifp/incidentify/user/service/UserServiceImpl.java
@@ -1,8 +1,14 @@
 package com.turing.alan.cpifp.incidentify.user.service;
 
+import org.springframework.beans.BeanUtils;
 import org.springframework.stereotype.Service;
+
+import java.util.List;
+
 import com.turing.alan.cpifp.incidentify.user.domain.UserEntity;
 import com.turing.alan.cpifp.incidentify.user.domain.UserRepository;
+import com.turing.alan.cpifp.incidentify.user.exception.UserAlreadyExistsException;
+import com.turing.alan.cpifp.incidentify.user.exception.UserDoesNotExistsException;
 
 @Service
 public class UserServiceImpl implements UserService {
@@ -13,30 +19,52 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public Iterable<UserEntity> getAll() {
-        return this.userRepository.findAll();
-        
+    public List<UserEntity> getAll() {
+        return (List<UserEntity>) this.userRepository.findAll();
     }
 
     @Override
     public UserEntity create(UserEntity entity) {
+
+        if (userRepository.existsByEmail(entity.getEmail())) {
+            throw new UserAlreadyExistsException();
+        }
         return this.userRepository.save(entity);
+    
     }
 
     @Override
     public UserEntity getOne(Long id) {
-        return this.userRepository.findById(id).orElseThrow();
+        return this.userRepository.findById(id).orElseThrow(
+            ()->new UserDoesNotExistsException()
+        );
     }
 
     @Override
     public void delete(Long id) {
-        this.userRepository.deleteById(id);
+        if (userRepository.existsById(id)) {
+            this.userRepository.deleteById(id);
+        }
+        else {
+            throw new UserDoesNotExistsException();
+        }
+        
     }
 
     @Override
-    public UserEntity update(UserEntity entity, Long id) {
+    public UserEntity update(UserEntity updatedEntity) {
         
-        return this.userRepository.save(entity);
+        long id = updatedEntity.getId();
+
+        UserEntity currentEntity = getOne(id);
+
+        // Copiamos todas las propiedades, excepto el id y el email
+        // que no son actualizables
+        BeanUtils.copyProperties(updatedEntity, 
+                                 currentEntity,
+                                 "id","email");
+        
+        return this.userRepository.save(currentEntity);
 
     }
